+++ crates/rslint_cli/src/files.rs
@@ -3,16 +3,12 @@
 use crate::lint_warn;
 use hashbrown::HashMap;
 use rayon::iter::{IntoParallelIterator, ParallelIterator};
+use rslint_core::File;
 use rslint_errors::file::{FileId, Files};
-use rslint_parser::{parse_with_syntax, FileKind, SyntaxNode};
 use std::fs::read_to_string;
 use std::ops::Range;
 use std::path::PathBuf;
-use std::sync::atomic::{AtomicUsize, Ordering};
 use walkdir::WalkDir;
-
-// 0 is reserved for "no file id" (virtual files)
-static FILE_ID_COUNTER: AtomicUsize = AtomicUsize::new(1);
 
 /// A list of ignored-by-default directory/file names
 const IGNORED: [&str; 1] = ["node_modules"];
@@ -24,7 +20,7 @@ const LINTED_FILES: [&str; 3] = ["js", "mjs", "ts"];
 // TODO: use IO_Uring for linux
 #[derive(Debug, Clone, PartialEq, Eq, Default)]
 pub struct FileWalker {
-    pub files: HashMap<usize, JsFile>,
+    pub files: HashMap<usize, File>,
 }
 
 impl Files for FileWalker {
@@ -69,7 +65,7 @@ impl FileWalker {
     }
 
     pub fn load_files(&mut self, paths: impl ParallelIterator<Item = PathBuf>) {
-        let jsfiles: HashMap<usize, JsFile> = paths
+        let jsfiles: HashMap<usize, File> = paths
             .filter(|p| {
                 !IGNORED.contains(&p.file_name().unwrap_or_default().to_string_lossy().as_ref())
             })
@@ -99,7 +95,7 @@ impl FileWalker {
                 };
                 Some((content, path.to_owned()))
             })
-            .map(|(src, path)| JsFile::new_concrete(src, path))
+            .map(|(src, path)| File::new_concrete(src, path))
             .map(|file| (file.id, file))
             .collect();
         self.files.extend(jsfiles);
@@ -125,97 +121,7 @@ impl FileWalker {
                 );
             };
             file.source = src;
-            file.line_starts = JsFile::line_starts(&file.source).collect();
+            file.line_starts = File::line_starts(&file.source).collect();
         }
-    }
-}
-
-/// A structure representing either a concrete (in-disk) or virtual (temporary/non-disk) js, ts, or mjs file.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct JsFile {
-    pub source: String,
-    /// The name of the file.
-    pub name: String,
-    /// The path in disk if this is a concrete file.
-    pub path: Option<PathBuf>,
-
-    /// The codespan id assigned to this file used to refer back to it.
-    pub id: usize,
-    /// The kind of file this is.
-    pub kind: FileKind,
-    /// The cached line start locations in this file.
-    pub line_starts: Vec<usize>,
-}
-
-impl JsFile {
-    pub fn new_concrete(source: String, path: PathBuf) -> Self {
-        let id = FILE_ID_COUNTER.fetch_add(1, Ordering::SeqCst);
-        let ext = path
-            .extension()
-            .map_or("".into(), |ext| ext.to_string_lossy());
-
-        let kind = match ext.as_ref() {
-            "mjs" => FileKind::Module,
-            "js" => FileKind::Script,
-            "ts" => FileKind::TypeScript,
-            _ => panic!("tried to make jsfile with extensions outside of `mjs`, `js`, or `ts`"),
-        };
-        let line_starts = Self::line_starts(&source).collect();
-
-        Self {
-            source,
-            name: path
-                .file_name()
-                .map_or(String::new(), |osstr| osstr.to_string_lossy().to_string()),
-            path: Some(path),
-            id,
-            kind,
-            line_starts,
-        }
-    }
-
-    pub fn update_src(&mut self, new: String) {
-        self.line_starts = Self::line_starts(&new).collect();
-        self.source = new;
-    }
-
-    // TODO: Needs to work correctly for \u2028, \u2029, and \r line endings
-    pub fn line_starts<'a>(source: &'a str) -> impl Iterator<Item = usize> + 'a {
-        std::iter::once(0).chain(source.match_indices('\n').map(|(i, _)| i + 1))
-    }
-
-    pub fn line_start(&self, line_index: usize) -> Option<usize> {
-        use std::cmp::Ordering;
-
-        match line_index.cmp(&self.line_starts.len()) {
-            Ordering::Less => self.line_starts.get(line_index).cloned(),
-            Ordering::Equal => Some(self.source.len()),
-            Ordering::Greater => None,
-        }
-    }
-
-    pub fn line_index(&self, byte_index: usize) -> usize {
-        match self.line_starts.binary_search(&byte_index) {
-            Ok(line) => line,
-            Err(next_line) => next_line - 1,
-        }
-    }
-
-    pub fn line_col_to_index(&self, line: usize, column: usize) -> Option<usize> {
-        let start = self.line_start(line)?;
-        Some(start + column)
-    }
-
-    fn line_range(&self, line_index: usize) -> Option<Range<usize>> {
-        let line_start = self.line_start(line_index)?;
-        let next_line_start = self.line_start(line_index + 1)?;
-
-        Some(line_start..next_line_start)
-    }
-
-    /// Parse this file into a syntax node, ignoring any errors produced. This
-    /// will use `parse_module` for `.mjs` and `parse_text` for `.js`
-    pub fn parse(&self) -> SyntaxNode {
-        parse_with_syntax(&self.source, self.id, self.kind.into()).syntax()
     }
 }