+++ crates/rslint_lsp/src/provider/diagnostics.rs
@@ -9,7 +9,6 @@ use rslint_core::{
     apply_top_level_directives, directives::DirectiveResult, run_rule, DirectiveParser,
 };
 use rslint_errors::{lsp::convert_to_lsp_diagnostic, Diagnostic as RslintDiagnostic};
-use rslint_parser::SyntaxNode;
 use std::{collections::HashMap, sync::Arc};
 use tower_lsp::lsp_types::*;
 
@@ -20,7 +19,7 @@ fn process_diagnostics(
     out: &mut Vec<Diagnostic>,
 ) {
     let files = document.files.clone();
-    let file_id = document.file_id;
+    let file_id = document.file.id;
 
     for diagnostic in diagnostics {
         if let Some(lsp_diag) = convert_to_lsp_diagnostic(
@@ -37,39 +36,33 @@ fn process_diagnostics(
 
 pub async fn publish_diagnostics(session: &Session, uri: Url) -> anyhow::Result<()> {
     let mut document = session.get_mut_document(&uri).await?;
-    let file_id = document.file_id;
 
     let mut new_store = session.store.clone();
     let DirectiveResult {
         directives,
         diagnostics: mut directive_diagnostics,
-    } = DirectiveParser::new_with_store(
-        SyntaxNode::new_root(document.parse.green()),
-        file_id,
-        &session.store,
-    )
-    .get_file_directives();
+    } = DirectiveParser::new_with_store(document.root.clone(), &document.file, &session.store)
+        .get_file_directives();
 
     apply_top_level_directives(
         directives.as_slice(),
         &mut new_store,
         &mut directive_diagnostics,
-        file_id,
+        document.file.id,
     );
 
     let verbose = false;
-    let src = Arc::from(document.text.clone());
+    let src = Arc::from(document.file.source.clone());
     let rule_results: HashMap<&str, rslint_core::RuleResult> = new_store
         .rules
         .par_iter()
         .map(|rule| {
-            let root = SyntaxNode::new_root(document.parse.green());
             (
                 rule.name(),
                 run_rule(
                     &**rule,
-                    file_id,
-                    root,
+                    document.file.id,
+                    document.root.clone(),
                     verbose,
                     &directives,
                     Arc::clone(&src),
@@ -93,7 +86,7 @@ pub async fn publish_diagnostics(session: &Session, uri: Url) -> anyhow::Result<
     process_diagnostics(
         &document,
         uri.clone(),
-        document.parse.parser_diagnostics().to_owned(),
+        document.parsing_errors.to_owned(),
         &mut diags,
     );